requirements.txt
djangorestframework>=3.11.0